introduce ERR_PTR

+++ kernel/inc/vfs/file.h
@@ -18,6 +18,6 @@ struct vfs_file *vfs_filedup(struct vfs_file *file);
 ssize_t vfs_fileread(struct vfs_file *file, void *buf, size_t n);
 int vfs_filestat(struct vfs_file *file, struct stat *stat);
 ssize_t vfs_filewrite(struct vfs_file *file, const void *buf, size_t n);
-int vfs_filelseek(struct vfs_file *file, loff_t offset, int whence, loff_t *new_pos);
+loff_t vfs_filelseek(struct vfs_file *file, loff_t offset, int whence, loff_t *new_pos);
 
 #endif // KERNEL_VIRTUAL_FILE_SYSTEM_FILE_H
